Output file name correctly generated
The previous version failed with input file names that contained no
path component (i.e. were absolutely relative to the current
directory). In this case, "" was joined with "/<input file base
name>.html", which gave a file in the root directory ("/"), which is
both wrong and potentially dangerous (moreover, "/" is normally not
writeable for normal users). Using the absolute path of the input file
avoids this.